requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
random-password-generator==2.2.0
sqlparse==0.4.4
tzdata==2023.3
